requirements.txt
python-fasthtml==0.2.1
uvicorn==0.30.1
beautifulsoup4==4.12.3
requests==2.32.3
